ditch unused concept of dead sockets
Way back in 2006, 23a909a21 introduced the concept of "unknown"
sockets supposedly alongside "dead" sockets.  The intention was
presumably to distinguish between them in some way, but in reality
it removed the code which ever marked a socket as dead.  So remove
this false distinction which was broken for 12 years.